changing check from id to name
