+++ src/components/pantry/Pantry.js
@@ -7,47 +7,18 @@ import Typography from '@mui/material/Typography';
 import PantryAutocomplete from '../PantryAutocomplete';
 import { ingredientList } from "../../../script/seedData";
 import Carousel from "react-elastic-carousel";
+import PantryList from './PantryList';
+import SearchWithPantry from '../RecipeResults/SearchWithPantry';
+import AdvancedSearch from '../RecipeResults/AdvancedSearch';
 
-//returns an object with all categories as keys, and a list of all ingredients in that category as the value
-const getAllCategories = (ingredients) => {
-  const mapCategoriesToIngredients = {};
-
-  ingredients.forEach(ingredient => {
-    const broadCategory = ingredient.broadCategory;
-    if(mapCategoriesToIngredients[broadCategory]){
-      mapCategoriesToIngredients[broadCategory].push(ingredient);
-    }
-    else{
-      mapCategoriesToIngredients[broadCategory] = [ingredient];
-    }
-  })
-
-  return mapCategoriesToIngredients;
-}
-
-
-export default function Pantry(){
-  const pantry = useSelector(state => state.selectedPantry);
-
-  if(!pantry.ingredients)
-    return null;
-
-  const categoriesWithIngredients = getAllCategories(pantry.ingredients);
-  const categories = Object.keys(categoriesWithIngredients).sort((a,b) => a.localeCompare(b));
-  
-  return(
-    <Box sx={{display:'flex', flexDirection:'column', alignItems:'center'}}>
-      <PantryAutocomplete selectedPantry={pantry} searchOptions={ingredientList} searchName='Add to Pantry' className="pantry-autocomplete"/>
-      {categories.map(category => {
-        return <Stack className='shelf' key={category} direction='row' sx={{flexWrap:'wrap', width:'400px', border: 'solid 1px black'}}>
-          <Typography variant='h5' sx={{flexBasis: '100%'}}>{category}</Typography>
-          <Carousel itemsToShow={4}>
-            {categoriesWithIngredients[category].map(ingredient =>{
-              return <PantryIcon ingredient={ingredient} key={ingredient.id}/>
-            })}
-          </Carousel>
-        </Stack>
-      })}
+export default function Pantry() {
+  return (
+    <Box sx={{display:'flex', justifyContent:'space-between'}}>
+      <PantryList/>
+      <Box sx={{display:'flex', flexDirection:'column', justifyContent:'center'}}>
+        <SearchWithPantry/>
+        <AdvancedSearch/>
+      </Box>
     </Box>
   )
 }
