+++ src/components/pantry/PantryDrawer.js
@@ -7,7 +7,7 @@ import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
-import SidePantry from './SidePantry';
+// import SidePantry from './SidePantry';
 import { Typography  } from "@mui/material";
 import PantrySelector from './PantrySelector';
 
@@ -51,7 +51,7 @@ export default function PantryDrawer({setPantryDrawerOpen, pantryIsOpen}) {
         </DrawerHeader>
         <Divider />
 
-        <SidePantry/>
+        {/* <SidePantry/> */}
       </Drawer>
     </Box>
   );
